Fix bug causing normal force min/max to be ignored.
This happened as the m_forceInequality was declared one entry too small. This caused the cl->forceConstraint in *acc-force.cpp to be too small as well and then ignore the last row. The last row defines the bound on the normal force min/max value.

This patch fixes the bug. It also adds extra assertions in the constraint-base to avoid these sort of bugs.