[notebooks] Fixing Waist Task
Adding and fixing the waist task for Pyrène.
The control is now working on both exercises.
Be careful to use joints and not frames when defining the Tasks and Trajectories. 
ie: use 'root_joint' and not 'base_link' for the waist.

Change the weights and gains: use the one of Pierre Fenrbach.