Add leak_rate parameter, change constructor & methods to take the contact
Instead of taking the name of the contact in the constructor, directly take the contact reference. Add setter/getter to set/get the associated contact.
Remove the useless setContactList method.
Change the first compute method (which takes the list of contacts in argument) so that it only checks if the associated contact is in the list.
Add bindings methods to set the gains and leak_rate parameters.
Add copyright institutions.