[Energy] Add python test and revert unnecessary changes
Remove blanks in the formulation file.
Fix comment in the task energy.

The python test creates a Romeo TsidBiped and add the energy task to the stack of task. Then the robot does a sinusoidal motion on its CoM.
If the CoM error does not diverge, the energy tank does not fall under its minimal value (0.1J) and the QP is solved then the test is ok.